Buggy proof of concept, big DRY exercise.
This more or less works for two directions, but not
perfectly and the code is massively replicated, with
minor differences.  Must write it once for all four
directions instead of 4x, which would be asking for
a plague of bugs.  Be sure to test random lost in space
pieces.  And implement an undo button.